LayoutTests/fast/canvas/canvas-createPattern-video-loading-expected.txt

Test the behavior of 2d canvas createPattern as a video is loaded

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.createElement('canvas').getContext('2d').createPattern(video, 'repeat') is null
PASS buffer[0] is within 2 of 200
PASS buffer[1] is within 2 of 200
PASS buffer[2] is within 2 of 0
PASS buffer[0] is within 2 of 200
PASS buffer[1] is within 2 of 200
PASS buffer[2] is within 2 of 0
PASS buffer[0] is within 2 of 200
PASS buffer[1] is within 2 of 200
PASS buffer[2] is within 2 of 0
PASS buffer[0] is within 2 of 200
PASS buffer[1] is within 2 of 200
PASS buffer[2] is within 2 of 0
PASS buffer[0] is within 2 of 57
PASS buffer[1] is within 2 of 202
PASS buffer[2] is within 2 of 203
PASS buffer[0] is within 2 of 57
PASS buffer[1] is within 2 of 202
PASS buffer[2] is within 2 of 203
PASS buffer[0] is within 2 of 57
PASS buffer[1] is within 2 of 202
PASS buffer[2] is within 2 of 203
PASS buffer[0] is within 2 of 57
PASS buffer[1] is within 2 of 202
PASS buffer[2] is within 2 of 203
PASS buffer[0] is within 2 of 53
PASS buffer[1] is within 2 of 203
PASS buffer[2] is within 2 of 9
PASS buffer[0] is within 2 of 53
PASS buffer[1] is within 2 of 203
PASS buffer[2] is within 2 of 9
PASS buffer[0] is within 2 of 53
PASS buffer[1] is within 2 of 203
PASS buffer[2] is within 2 of 9
PASS buffer[0] is within 2 of 53
PASS buffer[1] is within 2 of 203
PASS buffer[2] is within 2 of 9
PASS buffer[0] is within 2 of 200
PASS buffer[1] is within 2 of 200
PASS buffer[2] is within 2 of 0
PASS buffer[0] is within 2 of 200
PASS buffer[1] is within 2 of 200
PASS buffer[2] is within 2 of 0
PASS buffer[0] is within 2 of 200
PASS buffer[1] is within 2 of 200
PASS buffer[2] is within 2 of 0
PASS buffer[0] is within 2 of 200
PASS buffer[1] is within 2 of 200
PASS buffer[2] is within 2 of 0
PASS buffer[0] is within 2 of 57
PASS buffer[1] is within 2 of 202
PASS buffer[2] is within 2 of 203
PASS buffer[0] is within 2 of 57
PASS buffer[1] is within 2 of 202
PASS buffer[2] is within 2 of 203
PASS buffer[0] is within 2 of 57
PASS buffer[1] is within 2 of 202
PASS buffer[2] is within 2 of 203
PASS buffer[0] is within 2 of 57
PASS buffer[1] is within 2 of 202
PASS buffer[2] is within 2 of 203
PASS buffer[0] is within 2 of 53
PASS buffer[1] is within 2 of 203
PASS buffer[2] is within 2 of 9
PASS buffer[0] is within 2 of 53
PASS buffer[1] is within 2 of 203
PASS buffer[2] is within 2 of 9
PASS buffer[0] is within 2 of 53
PASS buffer[1] is within 2 of 203
PASS buffer[2] is within 2 of 9
PASS buffer[0] is within 2 of 53
PASS buffer[1] is within 2 of 203
PASS buffer[2] is within 2 of 9
PASS successfullyParsed is true

TEST COMPLETE
